Move button reset to style folder
