+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/molecules/Button.tsx
@@ -2,8 +2,8 @@ import React, { forwardRef } from 'react';
 import styled from 'styled-components';
 
 import { Colors, Radius, Spacings } from '../../foundations';
+import { buttonReset } from '../../styles';
 import { Flex } from '../layout';
-import { buttonReset } from '../mixins';
 import { BodySmallSemiBold } from '../typography';
 
 export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
